Split runtime tables into PSR-4 classes
