+++ components/com_breezingformsng/src/Table/RecordTable.php
@@ -1,0 +1,53 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vcmb\Component\BreezingformsNG\Site\Table;
+
+/**
+ * BreezingForms NG - A Joomla Forms Application
+ *
+ * @package BreezingFormsNG
+ * @copyright Copyright (C) 2008-2020 by Markus Bopp
+ * @copyright Copyright (C) 2024-2026 by XDA+GIL
+ * @license GNU General Public License version 2 or later; see LICENSE.txt
+ *
+ * SPDX-License-Identifier: GPL-2.0-or-later
+ */
+
+defined('_JEXEC') or die('Direct Access to this location is not allowed.');
+
+use Joomla\CMS\Table\Table;
+use Joomla\Database\DatabaseInterface;
+
+final class RecordTable extends Table
+{
+	public $id = null;     		// identifier
+	public $submitted = null;   // date and time
+	public $form = null;     	// form id
+	public $title = null;     	// form title
+	public $name = null;     	// form name
+	public $ip = null;     		// submitters ip
+	public $browser = null;     // browser
+	public $opsys = null;     	// operating system
+	public $provider = null;    // provider
+	public $viewed = null;     	// view status
+	public $exported = null;    // export status
+	public $archived = null;    // archive status
+	public $paypal_tx_id = null;
+	public $paypal_payment_date = null;
+	public $paypal_testaccount = null;
+	public $paypal_download_tries = null;
+
+	public function __construct(DatabaseInterface $db)
+	{
+		parent::__construct('#__facileforms_records', 'id', $db);
+	} // constructor
+
+	public function load($id = null, $reset = true)
+	{
+		return parent::load($id, $reset ?? true);
+	}
+
+} // class RecordTable
+
